more config from disk
If not specified in the install-config.yaml, use ~/.bluemix/apikey.json
for the apikey, and ~/.bluemix/config.json for the region, and only use
the env vars as a last check. And lots of informational messages because
this is a bit convoluded. We may decide to abandon the env vars but if
users are used to them, maybe we shouldn't.

Signed-off-by: Christy Norman <christy@linux.vnet.ibm.com>